# docs/swagger.yaml
basePath: /
definitions:
...
  controllers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  controllers.OTPResponse:
    properties:
      otp_url:
        type: string
    type: object
...
  models.Correo:
    properties:
      area:
        type: string
      created_at:
        type: string
      direccion:
        type: string
      dueno:
        $ref: '#/definitions/models.Persona'
      dueno_id:
        type: string
      id:
        type: string
      tipo:
        type: string
      updated_at:
        type: string
      userID:
        type: string
    type: object
...
  models.Direccion:
    properties:
      area:
        type: string
      created_at:
        type: string
      dueno:
        $ref: '#/definitions/models.Persona'
      dueno_id:
        type: string
      empleados:
        items:
          $ref: '#/definitions/models.Persona'
        type: array
      estado:
        type: string
      id:
        type: string
      lugar:
        type: string
      municipio:
        type: string
      nombre:
        type: string
      parroquia:
        type: string
      updated_at:
        type: string
      urbanizacion:
        type: string
      userID:
        type: string
      usuarios:
        items:
          $ref: '#/definitions/models.Persona'
        type: array
    type: object
...
  models.Empresa:
    properties:
      actividad_economica:
        type: string
      area:
        type: string
      created_at:
        type: string
      direccion:
        type: string
      empleados:
        items:
          $ref: '#/definitions/models.Persona'
        type: array
      id:
        type: string
      nombre:
        type: string
      representante:
        $ref: '#/definitions/models.Persona'
      representante_id:
        type: string
      rif:
        type: string
      socios:
        items:
          $ref: '#/definitions/models.Persona'
        type: array
      updated_at:
        type: string
      userID:
        type: string
    type: object
  models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  models.IIO:
    properties:
      area:
        type: string
      created_at:
        type: string
      descripcion:
        type: string
      fecha:
        type: string
      id:
        type: string
      imagen_url:
        type: string
      lugar:
        type: string
      modalidad:
        type: string
      nivel:
        type: string
      nombre:
        type: string
      parroquia:
        type: string
      redi:
        type: string
      relacion:
        items:
          $ref: '#/definitions/models.Persona'
        type: array
      tie:
        type: string
      updated_at:
        type: string
      urbanizacion:
        type: string
      userID:
        type: string
      verificada:
        type: boolean
      zodi:
        type: string
    type: object
...
  models.Nacionalidad:
    properties:
      area:
        type: string
      created_at:
        type: string
      gentilicio:
        type: string
      id:
        type: string
      updated_at:
        type: string
      userID:
        type: string
    type: object
  models.Pasaporte:
    properties:
      area:
        type: string
      codigo:
        type: string
      created_at:
        type: string
      foto:
        type: string
      id:
        type: string
      numero:
        type: string
      pais:
        type: string
      representante:
        $ref: '#/definitions/models.Persona'
      representante_id:
        type: string
      tipo:
        type: string
      updated_at:
        type: string
      userID:
        type: string
    type: object
  models.Persona:
    properties:
      alias:
        type: string
      apellido:
        type: string
      area:
        type: string
      cargo_actual:
        type: string
      cedula:
        type: string
      correo:
        type: string
      correos:
        items:
          $ref: '#/definitions/models.Correo'
        type: array
      created_at:
        type: string
      direcciones:
        items:
          $ref: '#/definitions/models.Direccion'
        type: array
      empresas:
        items:
          $ref: '#/definitions/models.Empresa'
        type: array
      estado_civil:
        type: string
      filiacion_politica:
        type: string
      id:
        type: string
      ideologia:
        type: string
      iio:
        items:
          $ref: '#/definitions/models.IIO'
        type: array
      informacion_de_interes:
        type: string
      nacionalidad:
        items:
          $ref: '#/definitions/models.Nacionalidad'
        type: array
      nombre:
        type: string
      pasaporte:
        items:
          $ref: '#/definitions/models.Pasaporte'
        type: array
      profesion:
        type: string
      redes:
        items:
          $ref: '#/definitions/models.Redes'
        type: array
      relacionados:
        items:
          $ref: '#/definitions/models.Persona'
        type: array
      religion:
        type: string
      telefono:
        type: string
      tipo_perfil:
        type: string
      updated_at:
        type: string
      userID:
        type: string
      valoraciones:
        type: string
      vehiculos:
        items:
          $ref: '#/definitions/models.Vehiculo'
        type: array
      visa:
        items:
          $ref: '#/definitions/models.Visa'
        type: array
    type: object
  models.Redes:
    properties:
      area:
        type: string
      created_at:
        type: string
      direccion:
        type: string
      dueno:
        $ref: '#/definitions/models.Persona'
      dueno_id:
        type: string
      id:
        type: string
      tipo:
        type: string
      updated_at:
        type: string
      userID:
        type: string
    type: object
...
  models.Vehiculo:
    properties:
      area:
        type: string
      año:
        type: string
      color:
        type: string
      created_at:
        type: string
      id:
        type: string
      linea:
        type: string
      marca:
        type: string
      matricula:
        type: string
      modelo:
        type: string
      numero:
        type: string
      propietario:
        items:
          $ref: '#/definitions/models.Persona'
        type: array
      tipo:
        type: string
      updated_at:
        type: string
      userID:
        type: string
      usuarios:
        items:
          $ref: '#/definitions/models.Persona'
        type: array
      vehiculo:
        items:
          $ref: '#/definitions/models.IIO'
        type: array
    type: object
  models.Visa:
    properties:
      aprobada:
        type: boolean
      area:
        type: string
      codigo:
        type: string
      created_at:
        type: string
      id:
        type: string
      pais:
        type: string
      representante:
        $ref: '#/definitions/models.Persona'
      representante_id:
        type: string
      tipo:
        type: string
      updated_at:
        type: string
      userID:
        type: string
      valoracion:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API SISTEMA DE EXPLORACION POPULAR.
  title: API
  version: "1.0"
paths:
...
  /iios:
    get:
      consumes:
      - application/json
      description: Obtiene todos los registros de IIO
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.IIO'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Obtiene todas las IIOs
      tags:
      - iio
    post:
      consumes:
      - multipart/form-data
      description: Crea un nuevo registro de IIO con los datos proporcionados
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Descripción del IIO
        in: formData
        name: descripcion
        required: true
        type: string
      - description: Fecha del IIO (YYYY-MM-DD)
        in: formData
        name: fecha
        required: true
        type: string
      - description: Lugar del IIO
        in: formData
        name: lugar
        required: true
        type: string
      - description: Modalidad del IIO
        in: formData
        name: modalidad
        required: true
        type: string
      - description: Nombre del IIO
        in: formData
        name: nombre
        required: true
        type: string
      - description: Parroquia del IIO
        in: formData
        name: parroquia
        required: true
        type: string
      - description: REDI del IIO
        in: formData
        name: redi
        required: true
        type: string
      - description: ZODI del IIO
        in: formData
        name: zodi
        required: true
        type: string
      - description: TIE del IIO
        in: formData
        name: tie
        required: true
        type: string
      - description: Área del IIO
        in: formData
        name: area
        required: true
        type: string
      - description: Imagen del IIO
        in: formData
        name: imagen
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.IIO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Crea un nuevo registro de IIO
      tags:
      - iio
  /iios/{id}:
    delete:
      consumes:
      - application/json
      description: Elimina un IIO por su ID
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID del IIO
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Elimina un IIO por su ID
      tags:
      - iio
    get:
      consumes:
      - application/json
      description: Obtiene los datos de un IIO por su ID
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID del IIO
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.IIO'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Obtiene un IIO por su ID
      tags:
      - iio
    put:
      consumes:
      - application/json
      description: Actualiza los datos de un IIO existente
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID del IIO
        in: path
        name: id
        required: true
        type: string
      - description: Datos del IIO actualizados
        in: body
        name: iio
        required: true
        schema:
          $ref: '#/definitions/models.IIO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.IIO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Actualiza un IIO existente
      tags:
      - iio
  /iios/filter:
    get:
      consumes:
      - application/json
      description: Obtiene registros de IIO en un periodo específico, y filtrados
        por tie, redi, zodi y modalidad
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Fecha de inicio (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: Fecha de fin (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      - description: TIE
        in: query
        name: tie
        type: string
      - description: REDI
        in: query
        name: redi
        type: string
      - description: ZODI
        in: query
        name: zodi
        type: string
      - description: Modalidad
        in: query
        name: modalidad
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.IIO'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Obtiene IIOs filtrados
      tags:
      - iio
...
  /users/{id}/otp-setup:
    get:
      consumes:
      - application/json
      description: Genera un secreto OTP y proporciona un URL en formato QR para configurar
        Google Authenticator
      parameters:
      - description: ID del usuario
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.OTPResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Genera un secreto OTP y un URL para configurar Google Authenticator
      tags:
      - OTP
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"